requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Audio processing and STT
openai-whisper==20231117
deepgram-sdk==3.2.7
sounddevice==0.4.6
numpy==1.26.3
scipy==1.12.0
librosa==0.10.1

# Async processing
aiofiles==23.2.1
httpx==0.26.0

# LLM integrations
openai==1.10.0
anthropic==0.9.0

# Template engine
jinja2==3.1.3

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.1
pyyaml==6.0.1
click==8.1.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
